+++ RenderSystems/GL3Plus/include/OgreGL3PlusPixelFormat.h
@@ -102,7 +102,7 @@ namespace Ogre {
             For compressed formats, constraints apply on minimum size and alignment
             so this might differ.
         */
-        static size_t getMaxMipmaps(size_t width, size_t height, size_t depth, PixelFormat format);
+        static uint32 getMaxMipmaps(uint32 width, uint32 height, uint32 depth, PixelFormat format);
     };
 }
